refactor(539): wire affectFileTreeSelection through the AffectedSettings checklist
Follows the AGENTS.md "Adding a New Setting (AffectedSettings)" steps the
first cut skipped: adds fileTreeSelection to IPeacockAffectedElementSettings,
getAffectedElements() and updateAffectedElements(), so the test harness
captures and restores it like every other affect* setting, and moves the
coverage into affected-elements.test.ts (enabled sets the three outline
tokens, fills stay untouched, disabled sets nothing). Folds the separate
collector into collectAccentBorderSettings, which already computes the same
accent color. Also moves the generic affected-settings screenshot back under
the Affected Elements intro so it no longer reads as part of the File Tree
Selection subsection.
