+++ src/configuration/read-configuration.ts
@@ -129,8 +129,6 @@ export function prepareColors(backgroundHex: string) {
 
   const statusBarSettings = collectStatusBarSettings(backgroundHex, keepForegroundColor);
 
-  const fileTreeSettings = collectFileTreeSettings(backgroundHex);
-
   // Merge all color settings
   const mergedSettings: ISettingsIndexer = {
     ...activityBarSettings,
@@ -139,7 +137,6 @@ export function prepareColors(backgroundHex: string) {
     ...accentBorderSettings,
     ...windowBorderSettings,
     ...squigglyBeGoneSettings,
-    ...fileTreeSettings,
   };
 
   const newColorCustomizations = sortSettingsIndexer(mergedSettings);
@@ -262,6 +259,7 @@ export function getAffectedElements() {
     tabActiveBorder: readConfiguration<boolean>(AffectedSettings.TabActiveBorder) || false,
     tabActiveBackground: readConfiguration<boolean>(AffectedSettings.TabActiveBackground) || false,
     windowBorder: readConfiguration<boolean>(AffectedSettings.WindowBorder) || false,
+    fileTreeSelection: readConfiguration<boolean>(AffectedSettings.FileTreeSelection) || false,
   } as IPeacockAffectedElementSettings;
 }
 
@@ -475,6 +473,16 @@ function collectAccentBorderSettings(backgroundHex: string) {
   if (isAffectedSettingSelected(AffectedSettings.TabActiveBackground)) {
     accentBorderSettings[ColorSettings.tabActiveBackground] = color;
   }
+  if (isAffectedSettingSelected(AffectedSettings.FileTreeSelection)) {
+    // Outline only (#539): a 1px ring in the accent color around the
+    // focused/selected row in the Explorer and other list/tree views. The
+    // theme's own selection and hover fills are left untouched so the change
+    // stays subtle. inactiveFocusOutline covers the common case where the
+    // editor has focus and the tree merely tracks the active file.
+    accentBorderSettings[ColorSettings.list_focusOutline] = color;
+    accentBorderSettings[ColorSettings.list_focusAndSelectionOutline] = color;
+    accentBorderSettings[ColorSettings.list_inactiveFocusOutline] = color;
+  }
   return accentBorderSettings;
 }
 
@@ -487,24 +495,6 @@ function collectWindowBorderSettings(backgroundHex: string) {
     windowBorderSettings[ColorSettings.window_inactiveBorder] = color;
   }
   return windowBorderSettings;
-}
-
-function collectFileTreeSettings(backgroundHex: string) {
-  const fileTreeSettings = {} as ISettingsIndexer;
-
-  if (isAffectedSettingSelected(AffectedSettings.FileTreeSelection)) {
-    // Outline only (#539): a 1px ring in the same accent the activity bar and
-    // accent borders use, drawn around the focused/selected row in the Explorer
-    // and other list/tree views. The theme's own selection and hover fills are
-    // left untouched so the change stays subtle. inactiveFocusOutline covers
-    // the common case where the editor has focus and the tree merely tracks
-    // the active file.
-    const { backgroundHex: color } = getElementStyle(backgroundHex, ElementNames.activityBar);
-    fileTreeSettings[ColorSettings.list_focusOutline] = color;
-    fileTreeSettings[ColorSettings.list_focusAndSelectionOutline] = color;
-    fileTreeSettings[ColorSettings.list_inactiveFocusOutline] = color;
-  }
-  return fileTreeSettings;
 }
 
 function collectSquigglyBeGoneSettings() {
