+++ src/models/interfaces.ts
@@ -44,6 +44,7 @@ export interface IPeacockAffectedElementSettings {
   tabActiveBorder: boolean;
   tabActiveBackground: boolean;
   windowBorder: boolean;
+  fileTreeSelection: boolean;
 }
 
 export interface IPeacockElementAdjustments {
